move radio node
