+++ packages/vlossom/src/components/vs-checkbox/VsCheckboxSet.vue
@@ -62,11 +62,10 @@ import { getInputProps, getInputOptionProps, getResponsiveProps } from '@/models
 import { VsComponent, VsNode, type ColorScheme } from '@/declaration';
 import { utils } from '@/utils';
 import VsInputWrapper from '@/components/vs-input-wrapper/VsInputWrapper.vue';
-import VsCheckboxNode from './VsCheckboxNode.vue';
-import { VsCheckboxNodeStyleSet } from './types';
+import VsCheckboxNode from '@/components/vs-checkbox/VsCheckboxNode.vue';
 import { useVsCheckboxSetRules } from './vs-checkbox-set-rules';
 
-import type { VsCheckboxSetStyleSet } from './types';
+import type { VsCheckboxNodeStyleSet, VsCheckboxSetStyleSet } from './types';
 
 const name = VsComponent.VsCheckboxSet;
 export default defineComponent({
