--- packages/vlossom/src/components/vs-radio-set/types.ts
@@ -1,6 +1,0 @@
-import type { VsRadioNodeStyleSet } from '@/nodes';
-
-export interface VsRadioSetStyleSet extends VsRadioNodeStyleSet {
-    radioGap?: string;
-    flexWrap?: string;
-}
