+++ packages/vlossom/src/components/vs-radio/VsRadioSet.vue
@@ -62,9 +62,9 @@ import { getInputProps, getInputOptionProps, getResponsiveProps } from '@/models
 import { VsComponent, VsNode, type ColorScheme } from '@/declaration';
 import { utils } from '@/utils';
 import VsInputWrapper from '@/components/vs-input-wrapper/VsInputWrapper.vue';
-import { VsRadioNode, VsRadioNodeStyleSet } from '@/nodes';
+import VsRadioNode from '@/components/vs-radio/VsRadioNode.vue';
 
-import type { VsRadioSetStyleSet } from './types';
+import type { VsRadioNodeStyleSet, VsRadioSetStyleSet } from './types';
 
 export default defineComponent({
     name: VsComponent.VsRadioSet,
