+++ packages/vlossom/src/components/vs-radio/types.ts
@@ -1,3 +1,16 @@
-import type { VsRadioNodeStyleSet } from '@/nodes';
+export interface VsRadioNodeStyleSet {
+    borderRadius?: string;
+    height?: string;
+    labelFontColor?: string;
+    labelFontSize?: string;
+    radioColor?: string;
+    radioSize?: string;
+    selectedLabelFontColor?: string;
+}
 
 export interface VsRadioStyleSet extends VsRadioNodeStyleSet {}
+
+export interface VsRadioSetStyleSet extends VsRadioNodeStyleSet {
+    radioGap?: string;
+    flexWrap?: string;
+}
