+++ packages/vlossom/src/nodes/index.ts
@@ -8,6 +8,3 @@ export { default as VsFocusTrap } from './vs-focus-trap/VsFocusTrap.vue';
 export type { VsModalNodeStyleSet, VsModalOptions } from './vs-modal-node/types';
 export { default as VsModalNode } from './vs-modal-node/VsModalNode.vue';
 export { default as VsModalView } from './vs-modal-node/VsModalView.vue';
-
-export type { VsRadioNodeStyleSet } from './vs-radio-node/types';
-export { default as VsRadioNode } from './vs-radio-node/VsRadioNode.vue';
