--- packages/vlossom/src/nodes/vs-radio-node/types.ts
@@ -1,9 +1,0 @@
-export interface VsRadioNodeStyleSet {
-    borderRadius?: string;
-    height?: string;
-    labelFontColor?: string;
-    labelFontSize?: string;
-    radioColor?: string;
-    radioSize?: string;
-    selectedLabelFontColor?: string;
-}
